Hide default supervisor tree to avoid race conditions
The user must not be able to modify default supervisor while
running. When declared as a global and exposed variable, this
cannot be prevented. Hiding the default supervisor to avoid
this problem. Should users need to use a custom supervisor,
they can use the zero value of supervisor.Supervisor as a starting
point.